# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DEV-001 Thyroid Pre-Visit Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /uploads/{appointmentId}:
    post:
      operationId: uploadDocument
      tags: [uploads]
      summary: Upload document metadata (patient auth)
      security:
        - patientAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"

    get:
      operationId: listDocuments
      tags: [uploads]
      summary: List patient's own documents (patient auth)
      security:
        - patientAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

  # ── Doctor clinical ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    patientAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Document:
      type: object
      required: [id, appointmentId, originalFileName, mimeType, fileSizeBytes, createdAt]
      properties:
        id:
          type: string
        appointmentId:
          type: string
        originalFileName:
          type: string
        mimeType:
          type: string
        fileSizeBytes:
          type: integer
        documentType:
          type: ["string", "null"]
        labStatus:
          type: ["string", "null"]
        deletedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    DocumentInput:
      type: object
      required: [originalFileName, mimeType, fileSizeBytes]
      properties:
        originalFileName:
          type: string
        mimeType:
          type: string
        fileSizeBytes:
          type: integer
        fileContentBase64:
          type: ["string", "null"]
        documentType:
          type: ["string", "null"]
        labStatus:
          type: ["string", "null"]
    UploadResult:
      type: object
      required: [document]
      properties:
        document:
          $ref: "#/components/schemas/Document"
        uploadUrl:
          type: ["string", "null"]
...