# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DEV-001 Thyroid Pre-Visit Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from './document';

export interface UploadResult {
  document: Document;
  /** @nullable */
  uploadUrl?: string | null;
}